# fix issue 17559 - [REG2.073.0] Wrong line number in stack trace