Do not generate debug line information for reuse of common sub expressions.

This also adds new syntax to d_do_test to enable OS specific command line arguments: if the test variable is followed by '(', a list of targets terminated by ')' can be specified to not apply the variable on other targets, e.g.:
```// REQUIRED_ARGS(linux freebsd): -L-export-dynamic```


~~This disables any optimizations in non-optimized builds. They badly effect the debug experience.~~

~~I measured the impact on build phobos unittests without optimizations (not the same system as for the demangling benchmarks):~~

|             | master | this PR
|---------|--------|--------
build time | 2min32s | 2min29s
exec time | 38.0s | 38.2s