@WalterBright 2017-06-28
I'm not too happy with this. Without comsubs and some jump optimization the generated code can be pretty awful. A better solution may be to not emit the line number for comsub'ed expressions.