@rainers 2017-06-29
I suspected it would be controversial, that's why I also didn't target the stable branch.

Line numbers are pretty bad with dmd, and you cannot really improve that if the optimizations throws the information away later. My hope is that we get a good step forward without these optimizations, but I have not tried it in a debugger yet.