@rainers 2017-06-29
The auto tester shows `runnable/test17559.d:30 _init [0x8048872]` in the call stack for posix systems. How do I get a `_Dmain` to show up there? This seems ok when built by circleCI.