@yazd 2017-06-29
I suspect the problem can be solved by removing this [line](https://github.com/dlang/dmd/blob/e785e39a0d65e8371e5352590950e64e1aa611ce/src/ddmd/e2ir.d#L1276) (untested). I can't think of a reason for why we need line info for referencing a symbol (if I understood the code correctly).
Actually, removing that line also resolves [this](https://issues.dlang.org/show_bug.cgi?id=15189). I started preparing a pull request but got stuck at writing the tests.