@rainers 2017-06-30
> I suspect the problem can be solved by removing this line (untested).

This looks quite different, I suspect the frontend attaches a bad line number to a symbol expression. This PR deals with common subexpressions being folded in the backend, thus removing the source line information of the duplicates.