@rainers 2017-06-30
> but I have not tried it in a debugger yet.

Tried that now, didn't look much better.

> A better solution may be to not emit the line number for comsub'ed expressions.

I tried to implement that now.