@rainers 2017-07-25
Ping! One of the few PRs that is all green ;)

@WalterBright the optimizations are back in place, the PR just removes the emission of debug info for duplicates of common sub expressions now.

This cannot easily target stable because the tests rely on testpdb.d being already there, but it's only on master. The bug is not really a regression, it was hidden in some situations before, though.