@braddr 2017-11-18
Manually merging this pull as it sat around long enough waiting to be marked approved that it accumulated github's max 1000 status updates per commit id and won't ever see more until a new commit becomes current for it.  Let that sink in... over 1000 builds done for a single pull request before it got marked for merging.